# api/openapi.yaml
openapi: 3.0.3
info:
  title: FolioPath HTTP API
  version: 0.1.0
  description: |
    Authoritative HTTP contract through the FolioPath POST-MVP-5 scope,
    revision 2. Revision-1 image semantic operations are contract-frozen.
    Revision-2 controlled tag suggestions, video semantic search, and local
    anonymous face/person curation remain gated from production composition
    until INT-S1R2 is accepted.

    All business endpoints use an authenticated same-origin browser session.
    State-changing authenticated operations additionally require the
    session-bound `X-CSRF-Token`. First-run setup and login do not yet have a
    session, so they require a validated same-origin `Origin` header.

    IDs and cursors are opaque. Media content endpoints accept only resource
    IDs, never filesystem paths. Management responses may expose a library
    root relative to the allowed `/library` container mount and a server-built
    `/library/...` display label; no response may expose a host path.
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
...
servers:
  - url: /
    description: Same-origin FolioPath server
security:
  - cookieAuth: []
...
paths:
...
  /api/v1/status:
    get:
      tags: [System]
      operationId: getSystemStatus
      summary: Get safe application status and capability flags
      description: |
        Requires an authenticated session. Values are safe for the application
        shell and never include database, cache, container-internal, or host paths.
      x-requirements: [FR-DEP-004, FR-AUTH-002, NFR-OPS-001, NFR-PRIV-001]
      responses:
        '200':
          description: Application status.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/ai/tag-vocabulary:
    get:
      tags: [Tag suggestions]
      operationId: getAITagVocabulary
      summary: Get the bounded controlled vocabulary used for suggestions
      x-requirements: [FR-INT-004, NFR-INT-004, NFR-INT-005]
      x-error-codes:
        '401': [authentication_required, session_expired]
        '429': [rate_limited]
        '500': [internal_error]
      responses:
        '200':
          description: Current immutable vocabulary snapshot projection.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AITagVocabulary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Tag suggestions]
      operationId: replaceAITagVocabulary
      summary: Publish a new controlled vocabulary snapshot
      security:
        - cookieAuth: []
          csrfToken: []
      x-requirements: [FR-INT-004, FR-INT-016, NFR-INT-003, NFR-INT-004]
      x-error-codes:
        '400': [invalid_request]
        '401': [authentication_required, session_expired]
        '403': [csrf_invalid]
        '404': [tag_not_found]
        '412': [precondition_failed]
        '422': [validation_failed]
        '428': [precondition_required]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/IfMatchHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AITagVocabularyUpdate'
      responses:
        '200':
          description: Newly published vocabulary snapshot.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AITagVocabulary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/libraries/{libraryId}/ai/tag-suggestions:
    parameters:
      - $ref: '#/components/parameters/LibraryIDParameter'
    get:
      tags: [Tag suggestions]
      operationId: listAITagSuggestions
      summary: List controlled tag suggestions and review decisions
      description: |
        Pending results use confidence DESC, suggestion ID ASC. Reviewed
        results use reviewedAt DESC, suggestion ID ASC. The opaque cursor
        binds library, status, tag, generation, vocabulary and progress.
      x-requirements: [FR-INT-004, FR-INT-015, NFR-INT-001, NFR-INT-004, NFR-INT-005]
      x-error-codes:
        '400': [invalid_request, invalid_cursor]
        '401': [authentication_required, session_expired]
        '404': [library_not_found, tag_not_found]
        '409': [suggestion_cursor_stale]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, accepted, dismissed]
            default: pending
        - name: tagId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ResourceID'
        - $ref: '#/components/parameters/CursorParameter'
        - $ref: '#/components/parameters/LimitParameter'
      responses:
        '200':
          description: Stable bounded suggestion page with truthful coverage.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AITagSuggestionPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/ai/tag-suggestion-reviews:
    post:
      tags: [Tag suggestions]
      operationId: reviewAITagSuggestions
      summary: Accept or dismiss at most 100 controlled suggestions
      security:
        - cookieAuth: []
          csrfToken: []
      x-requirements: [FR-INT-004, FR-INT-016, NFR-INT-003, NFR-INT-004]
      x-error-codes:
        '400': [invalid_request]
        '401': [authentication_required, session_expired]
        '403': [csrf_invalid]
        '409': [idempotency_conflict]
        '422': [validation_failed]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AITagSuggestionReviewRequest'
      responses:
        '200':
          description: Per-item outcomes; conflicts never create partial acceptance for an item.
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AITagSuggestionReviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/libraries/{libraryId}/ai/tag-suggestions/jobs:
    parameters:
      - $ref: '#/components/parameters/LibraryIDParameter'
    post:
      tags: [Tag suggestions]
      operationId: requestAITagSuggestionJob
      summary: Request missing-only or full controlled suggestion generation
      security:
        - cookieAuth: []
          csrfToken: []
      x-requirements: [FR-INT-004, FR-INT-016, NFR-INT-002]
      x-error-codes:
        '400': [invalid_request]
        '401': [authentication_required, session_expired]
        '403': [csrf_invalid]
        '404': [library_not_found]
        '409': [ai_disabled, model_unavailable, vocabulary_unavailable, idempotency_conflict]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SemanticJobRequest'
      responses:
        '202':
          description: Restart-safe bounded suggestion job accepted or coalesced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelOperation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/libraries/{libraryId}/ai/tag-suggestion-reviews/clear:
    parameters:
      - $ref: '#/components/parameters/LibraryIDParameter'
    post:
      tags: [Tag suggestions]
      operationId: clearAITagSuggestionReviews
      summary: Irreversibly clear accepted and dismissed suggestion-review records
      security:
        - cookieAuth: []
          csrfToken: []
      x-requirements: [FR-INT-004, NFR-INT-003]
      x-error-codes:
        '400': [invalid_request]
        '401': [authentication_required, session_expired]
        '403': [csrf_invalid]
        '404': [library_not_found]
        '409': [idempotency_conflict]
        '412': [precondition_failed]
        '428': [precondition_required]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/IfMatchHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClearTagSuggestionReviewsRequest'
      responses:
        '202':
          description: Review-clear operation accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelOperation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: foliopath_session
      description: |
        High-entropy, opaque session cookie backed by a revocable server-side
        session. It is HttpOnly and SameSite=Strict; deployments using HTTPS
        set Secure. The database stores no replayable plaintext token.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Session-bound token returned by setup, login, and session responses.
        It is required together with `cookieAuth` on authenticated
        state-changing operations. It must never be placed in a URL.
  parameters:
...
    LibraryIDParameter:
      name: libraryId
      in: path
      required: true
      description: Opaque media-library ID. It contains no path information.
      schema:
        $ref: '#/components/schemas/ResourceID'
...
    IfMatchHeader:
      name: If-Match
      in: header
      required: true
      description: Strong ETag from the latest GET. Missing values fail with 428 and stale values with 412.
      schema:
        type: string
        minLength: 3
        maxLength: 256
...
  headers:
...
    RequestID:
      description: Server-generated opaque correlation ID. Client-supplied values are not trusted.
      schema:
        type: string
        pattern: '^req_[A-Za-z0-9_-]+$'
        maxLength: 128
...
    ETag:
      description: |
        Strong opaque representation validator. For a Library it changes when
        any represented field changes, including scan-derived status/counts.
      schema:
        type: string
        example: '"fp-8b51cdd5"'
...
  responses:
    BadRequest:
      description: The request syntax, cursor, query combination, or encoded input is invalid.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: A valid administrator session is required or the session has expired.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: CSRF or same-origin validation failed. No protected action was executed.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested opaque resource ID does not identify a visible resource.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: A per-endpoint authentication, browse, scan, thumbnail, or media limit was reached.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected failure represented by a safe public code and request ID.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    Timestamp:
      type: string
      format: date-time
      description: UTC RFC 3339 timestamp.
...
    SystemStatus:
      type: object
      additionalProperties: false
      required:
        - version
        - apiVersion
        - runtimeState
        - initialized
        - readOnlyMedia
        - supportedLocales
        - supportedMedia
      properties:
        version:
          type: string
          minLength: 1
          maxLength: 128
          example: 0.1.0
        apiVersion:
          type: string
          enum: [v1]
        runtimeState:
          type: string
          enum: [ready, degraded]
          description: Degraded means the authenticated UI should inspect library and setting states.
        initialized:
          type: boolean
        readOnlyMedia:
          type: boolean
          enum: [true]
        supportedLocales:
          type: array
          minItems: 2
          uniqueItems: true
          items:
            type: string
            enum: [zh-CN, en]
        supportedMedia:
          $ref: '#/components/schemas/SupportedMedia'
...
    AITagVocabulary:
      type: object
      additionalProperties: false
      required: [revision, entries, publishedAt]
      properties:
        revision:
          type: integer
          format: int64
          minimum: 1
        entries:
          type: array
          maxItems: 1000
          uniqueItems: true
          items:
            $ref: '#/components/schemas/AITagVocabularyEntry'
        publishedAt:
          $ref: '#/components/schemas/Timestamp'
...
    AITagVocabularyUpdate:
      type: object
      additionalProperties: false
      required: [tagIds]
      properties:
        tagIds:
          type: array
          maxItems: 1000
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ResourceID'
...